Scene creation working
