+++ app/src/main/res/layout/fragment_scenes.xml
@@ -4,6 +4,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:id="@+id/sceneFragment"
     tools:context=".View.Fragments.SceneFragment">
 
     <include layout="@layout/frame_lists" />
